% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiply Two Numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
numeric
}
\description{
Multiply two numbers together.
}
\examples{
multiply(1,2)
}
