% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{Subtract Two Numbers}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
numeric
}
\description{
Subtracts two numbers.
}
\examples{
subtract(2,1)
}
